Consolidate binding code gen logic
