+++ compiler/src/main/java/com/squareup/anvil/compiler/codegen/ContributesMultibindingCodeGen.kt
@@ -9,14 +9,13 @@ import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.squareup.anvil.annotations.ContributesMultibinding
 import com.squareup.anvil.annotations.ContributesTo
-import com.squareup.anvil.annotations.internal.InternalBindingMarker
-import com.squareup.anvil.compiler.MULTIBINDING_MODULE_SUFFIX
 import com.squareup.anvil.compiler.api.AnvilApplicabilityChecker
 import com.squareup.anvil.compiler.api.AnvilContext
 import com.squareup.anvil.compiler.api.CodeGenerator
 import com.squareup.anvil.compiler.api.GeneratedFileWithSources
 import com.squareup.anvil.compiler.api.createGeneratedFile
 import com.squareup.anvil.compiler.checkNotGeneric
+import com.squareup.anvil.compiler.codegen.Contribution.Companion.generateFileSpec
 import com.squareup.anvil.compiler.codegen.ksp.AnvilSymbolProcessor
 import com.squareup.anvil.compiler.codegen.ksp.AnvilSymbolProcessorProvider
 import com.squareup.anvil.compiler.codegen.ksp.checkClassExtendsBoundType
@@ -33,35 +32,15 @@ import com.squareup.anvil.compiler.codegen.ksp.replaces
 import com.squareup.anvil.compiler.codegen.ksp.resolveBoundType
 import com.squareup.anvil.compiler.codegen.ksp.scope
 import com.squareup.anvil.compiler.contributesMultibindingFqName
-import com.squareup.anvil.compiler.internal.capitalize
-import com.squareup.anvil.compiler.internal.createAnvilSpec
 import com.squareup.anvil.compiler.internal.reference.asClassName
 import com.squareup.anvil.compiler.internal.reference.classAndInnerClassReferences
-import com.squareup.anvil.compiler.internal.reference.generateClassName
-import com.squareup.anvil.compiler.internal.reference.generateClassNameString
-import com.squareup.anvil.compiler.internal.safePackageString
 import com.squareup.anvil.compiler.qualifierKey
-import com.squareup.kotlinpoet.AnnotationSpec
-import com.squareup.kotlinpoet.ClassName
-import com.squareup.kotlinpoet.CodeBlock
-import com.squareup.kotlinpoet.FileSpec
-import com.squareup.kotlinpoet.FunSpec
-import com.squareup.kotlinpoet.KModifier
-import com.squareup.kotlinpoet.TypeSpec
-import com.squareup.kotlinpoet.asClassName
-import com.squareup.kotlinpoet.joinToCode
 import com.squareup.kotlinpoet.ksp.toAnnotationSpec
 import com.squareup.kotlinpoet.ksp.toClassName
 import com.squareup.kotlinpoet.ksp.writeTo
-import dagger.Binds
-import dagger.Module
-import dagger.Provides
-import dagger.multibindings.IntoMap
-import dagger.multibindings.IntoSet
 import org.jetbrains.kotlin.descriptors.ModuleDescriptor
 import org.jetbrains.kotlin.psi.KtFile
 import java.io.File
-import java.util.Comparator
 
 /**
  * Generates binding modules for every [ContributesMultibinding]-annotated class. If a class has repeated
@@ -69,112 +48,6 @@ import java.util.Comparator
  * annotated with [ContributesTo] for merging.
  */
 internal object ContributesMultibindingCodeGen : AnvilApplicabilityChecker {
-
-  private data class Contribution(
-    val origin: ClassName,
-    val scope: ClassName,
-    val isObject: Boolean,
-    val boundType: ClassName,
-    val replaces: List<ClassName>,
-    val qualifier: QualifierData?,
-    val mapKey: AnnotationSpec?,
-  ) {
-    data class QualifierData(
-      val annotationSpec: AnnotationSpec,
-      val key: String,
-    )
-    companion object {
-      val COMPARATOR: Comparator<Contribution> = compareBy<Contribution> { it.scope.canonicalName }
-        .thenComparing(compareBy { it.boundType.canonicalName })
-        .thenComparing(compareBy { it.replaces.joinToString(transform = ClassName::canonicalName) })
-    }
-  }
-
-  private fun generate(
-    className: ClassName,
-    contributions: List<Contribution>,
-  ): FileSpec {
-    val fileName = className.generateClassName(suffix = MULTIBINDING_MODULE_SUFFIX).simpleName
-    val generatedPackage = className.packageName.safePackageString(dotPrefix = true)
-
-    val specs = contributions.map { contribution ->
-      // Combination name of origin, scope, and boundType
-      val suffix = "As" +
-        contribution.boundType.generateClassNameString().capitalize() +
-        "To" +
-        contribution.scope.generateClassNameString().capitalize() +
-        "MultiBindingModule"
-
-      val contributionName =
-        className.generateClassName(suffix = suffix).simpleName
-      val builder = if (contribution.isObject) {
-        TypeSpec.objectBuilder(contributionName)
-      } else {
-        TypeSpec.interfaceBuilder(contributionName)
-      }
-      builder.apply {
-        addAnnotation(Module::class)
-        addAnnotation(
-          AnnotationSpec.builder(ContributesTo::class)
-            .addMember("scope = %T::class", contribution.scope)
-            .apply {
-              if (contribution.replaces.isNotEmpty()) {
-                addMember(
-                  "replaces = %L",
-                  contribution.replaces.map { CodeBlock.of("%T::class", it) }
-                    .joinToCode(prefix = "[", suffix = "]"),
-                )
-              }
-            }
-            .build(),
-        )
-
-        addAnnotation(
-          AnnotationSpec.builder(
-            InternalBindingMarker::class.asClassName(),
-          )
-            .addMember("originClass = %T::class", contribution.origin)
-            .addMember("isMultibinding = true")
-            .apply {
-              contribution.qualifier?.key?.let { qualifierKey ->
-                addMember("qualifierKey = %S", qualifierKey)
-              }
-            }
-            .build(),
-        )
-
-        val nameSuffix = contribution.boundType.simpleName.capitalize()
-        val functionBuilder = if (contribution.isObject) {
-          FunSpec.builder("provide$nameSuffix")
-            .addAnnotation(Provides::class)
-            .addStatement("return %T", className)
-        } else {
-          FunSpec.builder("bind$nameSuffix")
-            .addModifiers(KModifier.ABSTRACT)
-            .addAnnotation(Binds::class)
-            .addParameter("real", className)
-        }
-
-        addFunction(
-          functionBuilder
-            .apply {
-              if (contribution.mapKey == null) {
-                addAnnotation(IntoSet::class)
-              } else {
-                addAnnotation(IntoMap::class)
-                addAnnotation(contribution.mapKey)
-              }
-              contribution.qualifier?.let { addAnnotation(it.annotationSpec) }
-            }
-            .returns(contribution.boundType)
-            .build(),
-        )
-      }.build()
-    }
-    return FileSpec.createAnvilSpec(generatedPackage, fileName) {
-      addTypes(specs.sortedBy { it.name })
-    }
-  }
 
   override fun isApplicable(context: AnvilContext) = !context.generateFactoriesOnly
 
@@ -202,7 +75,6 @@ internal object ContributesMultibindingCodeGen : AnvilApplicabilityChecker {
           clazz.checkClassExtendsBoundType(contributesMultibindingFqName, resolver)
 
           // All good, generate away
-          val className = clazz.toClassName()
           val contributions = clazz.getKSAnnotationsByType(ContributesMultibinding::class)
             .toList()
             .also { it.checkNoDuplicateScopeAndBoundType(clazz) }
@@ -226,7 +98,7 @@ internal object ContributesMultibindingCodeGen : AnvilApplicabilityChecker {
                 .filter { it.isMapKey() }
                 .singleOrNull()
                 ?.toAnnotationSpec()
-              Contribution(
+              Contribution.MultiBinding(
                 clazz.toClassName(),
                 scope,
                 clazz.classKind == ClassKind.OBJECT,
@@ -236,11 +108,8 @@ internal object ContributesMultibindingCodeGen : AnvilApplicabilityChecker {
                 mapKey,
               )
             }
-            .distinct()
-            // Give it a stable sort.
-            .sortedWith(Contribution.COMPARATOR)
 
-          generate(className, contributions)
+          contributions.generateFileSpec()
             .writeTo(
               codeGenerator = env.codeGenerator,
               aggregating = false,
@@ -280,8 +149,7 @@ internal object ContributesMultibindingCodeGen : AnvilApplicabilityChecker {
           clazz.checkClassExtendsBoundType(contributesMultibindingFqName)
         }
         .map { clazz ->
-          val className = clazz.asClassName()
-          val scopes = clazz.annotations
+          val contributions = clazz.annotations
             .find(contributesMultibindingFqName)
             .also { it.checkNoDuplicateScopeAndBoundType() }
             .map {
@@ -304,7 +172,7 @@ internal object ContributesMultibindingCodeGen : AnvilApplicabilityChecker {
               val mapKey = clazz.annotations
                 .find { it.isMapKey() }
                 ?.toAnnotationSpec()
-              Contribution(
+              Contribution.MultiBinding(
                 clazz.asClassName(),
                 scope,
                 clazz.isObject(),
@@ -314,11 +182,8 @@ internal object ContributesMultibindingCodeGen : AnvilApplicabilityChecker {
                 mapKey,
               )
             }
-            .distinct()
-            // Give it a stable sort.
-            .sortedWith(Contribution.COMPARATOR)
 
-          val spec = generate(className, scopes)
+          val spec = contributions.generateFileSpec()
 
           createGeneratedFile(
             codeGenDir = codeGenDir,
